+++ bin/replicante/src/components/core_api/apply/appliers.rs
@@ -11,6 +11,7 @@ use super::agent_action;
 use super::discovery_settings;
 use super::namespace;
 use super::orchestrator_action;
+use super::platform;
 use crate::Result;
 
 const APIV_REPLI_V0: &str = "replicante.io/v0";
@@ -19,6 +20,7 @@ const KIND_DISCOVERY_SETTING: &str = "DiscoverySettings";
 const KIND_NAMESPACE: &str = "Namespace";
 const KIND_NODE_ACTION: &str = "NodeAction";
 const KIND_ORCHESTRATOR_ACTION: &str = "OrchestratorAction";
+const KIND_PLATFORM: &str = "Platform";
 
 /// Type of closure that handles a specific `kind` for a specific `apiVersion`.
 pub type Applier = Box<dyn Fn(ApplierArgs) -> Result<Value>>;
@@ -48,6 +50,9 @@ pub fn find(object: &ApplyObject) -> Option<Applier> {
         }
         (APIV_REPLI_V0, KIND_ORCHESTRATOR_ACTION) => {
             Some(Box::new(orchestrator_action::replicante_io_v0))
+        }
+        (APIV_REPLI_V0, KIND_PLATFORM) => {
+            Some(Box::new(platform::replicante_io_v0))
         }
         _ => None,
     }