+++ bin/replicante/src/components/core_api/apply/mod.rs
@@ -30,6 +30,7 @@ mod discovery_settings;
 mod metrics;
 mod namespace;
 mod orchestrator_action;
+mod platform;
 mod validate;
 
 pub use metrics::register_metrics;