+++ models/core/src/events/platform.rs
@@ -1,3 +1,4 @@
+use replisdk::core::models::platform::Platform;
 use serde::Deserialize;
 use serde::Serialize;
 
@@ -5,47 +6,46 @@ use super::Event;
 use super::EventBuilder;
 use super::Payload;
 use crate::scope::EntityId;
-use crate::scope::Namespace;
 
-/// Enumerates all possible namespace events emitted by the system.
+/// Enumerates all possible platform events emitted by the system.
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 #[serde(tag = "event", content = "payload")]
 // TODO: use when possible #[non_exhaustive]
-pub enum NamespaceEvent {
-    /// A Namespace object was applied.
+pub enum PlatformEvent {
+    /// A [`Platform`] object was applied.
     ///
     /// This event is emitted even if the object already exists and was not changed.
-    #[serde(rename = "NAMESPACE_APPLY")]
-    Apply(Namespace),
+    #[serde(rename = "PLATFORM_APPLY")]
+    Apply(Platform),
 }
 
-impl NamespaceEvent {
+impl PlatformEvent {
     /// Returns the event "code", the string that represents the event type.
     pub fn code(&self) -> &'static str {
         match self {
-            NamespaceEvent::Apply(_) => "NAMESPACE_APPLY",
+            PlatformEvent::Apply(_) => "PLATFORM_APPLY",
         }
     }
 
     /// Identifier of the namespace.
     pub fn entity_id(&self) -> EntityId {
         let namespace = match self {
-            NamespaceEvent::Apply(ns) => &ns.ns_id,
+            PlatformEvent::Apply(platform) => &platform.ns_id,
         };
         EntityId::Namespace(namespace)
     }
 }
 
-/// Build `NamespaceEventBuilder`s, validating inputs.
-pub struct NamespaceEventBuilder {
+/// Build `PlatformEvent`s, validating inputs.
+pub struct PlatformEventBuilder {
     pub(super) builder: EventBuilder,
 }
 
-impl NamespaceEventBuilder {
-    /// Build a `NamespaceEventEvent::Apply` event.
-    pub fn apply(self, namespace: Namespace) -> Event {
-        let event = NamespaceEvent::Apply(namespace);
-        let payload = Payload::Namespace(event);
+impl PlatformEventBuilder {
+    /// Build a `PlatformEvent::Apply` event.
+    pub fn apply(self, platform: Platform) -> Event {
+        let event = PlatformEvent::Apply(platform);
+        let payload = Payload::Platform(event);
         self.builder.finish(payload)
     }
 }